# docs/flatmates-openapi.yaml
openapi: 3.1.0
info:
  title: 360 Flatmates API
  description: |
    Complete API for the 360 Flatmates web platform — a swipe-based flatmate-finding app for India's young professionals.

    ## Authentication
    All endpoints require authentication via Bearer token in the `Authorization` header, unless marked as public.
    Tokens are obtained via Supabase Auth (Phone OTP + Password).

    ## User Modes
    - **room_poster** — Has a flat, wants to rent out a spare room
    - **co_hunter** — Looking for people to flat-search with
    - **open_to_both** — Flexible, open to either

    ## Base URL
    `/api/v1`
  version: 1.0.0
  contact:
    name: 360 Ghar
    url: https://360ghar.com
  license:
    name: Confidential
    url: https://360ghar.com
servers:
  - url: https://api.360ghar.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local Development
...
paths:
...
  /flatmates/profile:
    get:
      tags: [Profiles]
      summary: Get my flatmate profile
      description: Fetch the authenticated user's flatmate profile.
      operationId: getMyFlatmatesProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Flatmate profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlatmatesProfile'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Profiles]
      summary: Update my flatmate profile
      description: Update the authenticated user's flatmate profile fields.
      operationId: updateMyFlatmatesProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlatmatesProfileUpdate'
      responses:
        '200':
          description: Updated flatmate profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlatmatesProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /flatmates/moderation/listings:
    get:
      tags: [Admin Moderation]
      summary: Get listings for moderation
      description: |
        Get listings pending moderation review. Requires admin role.
        Auto-pauses expired listings before returning results.
      operationId: getPendingListings
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            default: pending_review
          description: Filter by moderation status
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Paginated listings for moderation
          content:
            application/json:
              schema:
                type: object
                properties:
                  listings:
                    type: array
                    items:
                      $ref: '#/components/schemas/FlatmateListingAdmin'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '403':
          $ref: '#/components/responses/Forbidden'
  /flatmates/moderation/listings/{listing_id}:
    put:
      tags: [Admin Moderation]
      summary: Moderate a listing
      description: |
        Moderate a listing: approve, reject, or request edit. Requires admin role.
        On first approval, a 24-hour launch boost is automatically applied.
      operationId: moderateListing
      security:
        - BearerAuth: []
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingModerationAction'
      responses:
        '200':
          description: Moderation action applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  listing_id:
                    type: integer
                  action:
                    type: string
                    enum: [approve, reject, request_edit]
                  status:
                    type: string
                  reason:
                    type: string
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /flatmates/moderation/reports:
    get:
      tags: [Admin Moderation]
      summary: Get reports for moderation
      description: Get user reports pending review. Requires admin role.
      operationId: getPendingReports
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            default: open
          description: Filter by report status
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Paginated reports for moderation
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReportAdmin'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '403':
          $ref: '#/components/responses/Forbidden'
  /flatmates/moderation/reports/{report_id}:
    put:
      tags: [Admin Moderation]
      summary: Moderate a report
      description: |
        Moderate a user report: dismiss, warn user, suspend user, or escalate. Requires admin role.
        Suspension deactivates the reported user's account.
      operationId: moderateReport
      security:
        - BearerAuth: []
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportModerationAction'
      responses:
        '200':
          description: Moderation action applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_id:
                    type: integer
                  action:
                    type: string
                    enum: [dismiss, warn_user, suspend_user, escalate]
                  status:
                    type: string
                  notes:
                    type: string
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /flatmates/moderation/prescreen/{listing_id}:
    post:
      tags: [Admin Moderation]
      summary: Pre-screen a listing
      description: |
        Run deterministic AI pre-screening for a flatmates listing.
        Checks: photos present, required fields, pricing anomalies, spam keywords.
        Requires admin role.
      operationId: prescreenListing
      security:
        - BearerAuth: []
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pre-screen result
          content:
            application/json:
              schema:
                type: object
                properties:
                  listing_id:
                    type: integer
                  prescreen_result:
                    type: object
                  prescreen_flags:
                    type: array
                    items:
                      type: string
                  prescreen_reason:
                    type: string
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===================== WEB DISCOVERY =====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    FlatmatesProfile:
      type: object
      properties:
        id:
          type: integer
        full_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        profile_image_url:
          type: string
          format: uri
        mode:
          $ref: '#/components/schemas/FlatmatesMode'
        profile_status:
          $ref: '#/components/schemas/FlatmatesProfileStatus'
        onboarding_completed:
          type: boolean
          default: false
        bio:
          type: string
          maxLength: 500
        age:
          type: integer
          minimum: 18
          maximum: 100
        profession:
          type: string
        budget_min:
          type: number
          minimum: 0
        budget_max:
          type: number
          minimum: 0
        move_in_timeline:
          type: string
          enum: [immediate, this_month, next_month, flexible]
        city:
          type: string
        locality:
          type: string
        sleep_schedule:
          type: string
          enum: [early_bird, flexible, night_owl]
        cleanliness:
          type: string
          enum: [minimal, tidy, spotless]
        food_habits:
          type: string
          enum: [vegetarian, non_vegetarian, eggetarian, no_preference]
        smoking_drinking:
          type: string
          enum: [neither, smoke_outside, no_preference]
        guests_policy:
          type: string
          enum: [no_overnight, occasional_ok, open_house]
        work_style:
          type: string
          enum: [wfh_mostly, office_mostly, mixed]
        gender:
          type: string
        gender_preference:
          type: string
          enum: [male, female, any]
        preferences:
          type: object
          additionalProperties: true
          description: Extended flatmate preferences (JSONB)
        last_active_at:
          type: string
          format: date-time
    FlatmatesProfileUpdate:
      type: object
      properties:
        full_name:
          type: string
        profile_image_url:
          type: string
          format: uri
        mode:
          $ref: '#/components/schemas/FlatmatesMode'
        profile_status:
          $ref: '#/components/schemas/FlatmatesProfileStatus'
        onboarding_completed:
          type: boolean
        bio:
          type: string
          maxLength: 500
        age:
          type: integer
          minimum: 18
          maximum: 100
        profession:
          type: string
        budget_min:
          type: number
          minimum: 0
        budget_max:
          type: number
          minimum: 0
        move_in_timeline:
          type: string
          enum: [immediate, this_month, next_month, flexible]
        city:
          type: string
        locality:
          type: string
        sleep_schedule:
          type: string
          enum: [early_bird, flexible, night_owl]
        cleanliness:
          type: string
          enum: [minimal, tidy, spotless]
        food_habits:
          type: string
          enum: [vegetarian, non_vegetarian, eggetarian, no_preference]
        smoking_drinking:
          type: string
          enum: [neither, smoke_outside, no_preference]
        guests_policy:
          type: string
          enum: [no_overnight, occasional_ok, open_house]
        work_style:
          type: string
          enum: [wfh_mostly, office_mostly, mixed]
        gender:
          type: string
        gender_preference:
          type: string
          enum: [male, female, any]
        preferences:
          type: object
          additionalProperties: true
...
    FlatmateListingAdmin:
      type: object
      description: Listing representation for admin moderation queue
      properties:
        id:
          type: integer
        title:
          type: string
        owner_id:
          type: integer
        owner_name:
          type: string
        owner_phone:
          type: string
        city:
          type: string
        locality:
          type: string
        monthly_rent:
          type: number
        main_image_url:
          type: string
          format: uri
        image_urls:
          type: array
          items:
            type: string
            format: uri
        description:
          type: string
        moderation_status:
          type: string
        created_at:
          type: string
          format: date-time
        ai_prescreen_result:
          type: object
        ai_prescreen_flags:
          type: array
          items:
            type: string
        ai_prescreen_reason:
          type: string

    # ===================== NOTIFICATION SCHEMAS =====================
...
    ReportAdmin:
      type: object
      properties:
        id:
          type: integer
        reporter_user_id:
          type: integer
        reporter_name:
          type: string
        reported_user_id:
          type: integer
        reported_name:
          type: string
        conversation_id:
          type: integer
          nullable: true
        property_id:
          type: integer
          nullable: true
        reason:
          $ref: '#/components/schemas/UserReportReason'
        status:
          $ref: '#/components/schemas/UserReportStatus'
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ===================== INTERACTION SCHEMAS =====================
...
    ListingModerationAction:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [approve, reject, request_edit]
        reason:
          type: string
          maxLength: 500
    ReportModerationAction:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [dismiss, warn_user, suspend_user, escalate]
        notes:
          type: string
          maxLength: 500

    # ===================== WEB DISCOVERY SCHEMAS =====================
...
    FlatmatesMode:
      type: string
      enum: [room_poster, co_hunter, open_to_both]
    FlatmatesProfileStatus:
      type: string
      enum: [draft, pending_review, active, paused]
...
    UserReportReason:
      type: string
      enum: [spam, fake_profile, abuse, inappropriate, other]
    UserReportStatus:
      type: string
      enum: [open, reviewed, dismissed, actioned]
...
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
        error_code:
          type: string
          description: Optional machine-readable error code